Dep-eject api.fullscreen
