+++ lib/depject/index.js
@@ -12,7 +12,6 @@ module.exports = {
       sync: require('./about/sync.js')
     },
     app: {
-      fullscreen: require('./app/fullscreen.js'),
       html: {
         'progress-notifier': require('./app/html/progress-notifier.js'),
         search: require('./app/html/search.js')
