+++ lib/main-window.js
@@ -16,6 +16,7 @@ const requireStyle = require('require-style')
 const ssbUri = require('ssb-uri')
 const pull = require('pull-stream')
 const moment = require('moment-timezone')
+const fullscreen = require('./fullscreen.js')
 
 const localTimezone = moment.tz.guess()
 moment.tz.setDefault(localTimezone)
@@ -38,7 +39,6 @@ module.exports = function (config) {
     'app.html.search': 'first',
     'app.html.channels': 'first',
     'app.views': 'first',
-    'app.fullscreen': 'first',
     'app.sync.externalHandler': 'first',
     'app.html.progressNotifier': 'first',
     'about.async.latestValues': 'first',
@@ -194,7 +194,7 @@ module.exports = function (config) {
   )
 
   const container = h(`MainWindow -${process.platform}`, {
-    classList: [when(api.app.fullscreen(), '-fullscreen')],
+    classList: [when(fullscreen(), '-fullscreen')],
     'ev-dragover': preventDefault,
     'ev-drop': preventDefault,
     'ev-dragstart': preventDefaultUnlessImage
